% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvi_dglm.R
\name{nvi_dglm}
\alias{nvi_dglm}
\title{nvi_dglm: DGLM for multivariate binominal data.}
\usage{
nvi_dglm(Data, GG, FF, m0, C0, W = NULL, delta = 0.75)
}
\arguments{
\item{Data}{[\code{data.frame}]. Data frame with three elements:\cr
\code{t}: Vector with times/ dates.\cr
\code{n}: Matrix, where each row represents the number at risk for different
levels at associated time step.\cr
\code{y}: Matrix, where each row represents the number of incidents for different
levels at associated time step.}

\item{GG}{[\code{numeric matrix}] System matrix G. Assumed constant for all time elements.}

\item{FF}{[\code{numeric matrix}] Design matrix F. Assumed constant for all time elements.}

\item{m0}{[\code{numeric vector}] Prior mean for (latent) parameter vector \eqn{\theta}.}

\item{C0}{[\code{numeric matrix}] Prior variance for (latent) parameter vector \eqn{\theta}.}

\item{W}{[\code{numeric matrix}] System variance.}

\item{delta}{[\code{numeric scalar}]. Discount factor. Only used if \code{W} is \code{NULL}.}
}
\value{
A list with elements at, Rt,mt,Ct,ft,Qt,pt,etaHat,vHat,Qst
}
\description{
The function \code{DGLM_Multivariate_Binominal} fits a Dynamic Generalized
Linear Model (DGLM) to the data input. See "Bono, C., Cornou, C., Lundbye-Christensen, S., 
& Kristensen, A. R. (2013). Dynamic production monitoring in pig herds II. 
Modeling and monitoring farrowing rate at herd level. Livestock Science, 155(1), 92-102.
Doi: 10.1016/j.livsci.2013.03.026" for details.
}
\details{
This code is not yet implemented with the "formula option", ref. 
\link[NVI_dynamic_linear_models]{nvi_dlm}. The code might be merged to one
general function covering both DLM and DGLM as much of the framework is the same.
}
\examples{
\dontrun{

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
