% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvi_dlm.R
\name{nvi_dlm}
\alias{nvi_dlm}
\title{nvi_dlm: DLM for multivariate normal data.}
\usage{
nvi_dlm(
  systemform = NULL,
  errorform = NULL,
  varcomps = list(main = 1, trend = 10^-3, error = list(main = 1, a = 0.1, b = 0.1),
    waves = 10^-3, a = NULL, b = NULL, dist = list(kappa = 10, nu = 1/5, sigma_d2 = 1)),
  data,
  GG = NULL,
  FF = NULL,
  m0 = NULL,
  C0 = NULL,
  W = NULL,
  V = NULL,
  delta = 0.75,
  smoother_run = FALSE,
  isnullw = FALSE
)
}
\arguments{
\item{systemform}{[\code{formula}]. A object of class "formula", see 
\link[stats]{formula}. If not \code{NULL} passed directly to 
function \link[NVI_dynamic_linear_models]{nvi_dlm_design}, together 
with \code{errorform} and \code{varcomps}. See 
\link[NVI_dynamic_linear_models]{nvi_dlm_design} for details.}

\item{errorform}{[\code{formula}]. A object of class "formula", see 
\link[stats]{formula}. If not \code{NULL} passed directly to 
function \link[NVI_dynamic_linear_models]{nvi_dlm_design}, 
together with \code{systemform} and \code{varcomps}. See 
\link[NVI_dynamic_linear_models]{nvi_dlm_design} for details.}

\item{data}{[\code{data.frame}]. List with at least two elements:\cr
\code{t}: Vector with times/ dates.\cr
\code{y}: Matrix, of size \code{N x q} where each row represents the 
observations for \code{q} different sites/ levels at associated time step. 
(\code{N} timesteps in total). Might be given name in accordance with
\code{systemform}. 
If the "formula" option is used, \code{data} must also contain elements 
defining \code{nested factors}(of length \code{q}), \code{wavelengths} defining
harmonic waves, \code{distance matrix} (of size \code{q x 2}) associated with elements included
in \code{systemform} and \code{errorform} respectively.}

\item{GG}{[\code{numeric matrix}] System matrix G. 
Assumed constant for all time elements. 
Not used if "formula option is applied"}

\item{FF}{[\code{numeric matrix}] Design matrix F. 
Assumed constant for all time elements. 
Not used if "formula option is applied"}

\item{m0}{[\code{numeric vector}] Prior mean for (latent) 
parameter vector \eqn{\theta}.}

\item{C0}{[\code{numeric matrix}] Prior variance for (latent) 
parameter vector \eqn{\theta}.}

\item{W}{[\code{numeric matrix}] System variance.
Assumed constant for all time elements. 
Not used if "formula option is applied"}

\item{V}{[\code{numeric matrix}] Error variance.
Assumed constant for all time elements. 
Not used if "formula option is applied"}

\item{delta}{[\code{numeric scalar}]. Discount factor. 
If not \code{NULL} discount factor is applied and input argument 
\code{W} (system variance) is ignored, also when "formula option" is
applied. \code{delta} might be given as a matrix of size \code{p x p}}

\item{smoother_run}{[\code{logical}]. If the DLM smoother is to 
be run}

\item{isnullw}{[\code{logical}]. If the system variance is to be returned as 
\code{NULL}. Must be set till \code{TRUE} if discount factor method is to 
be applied in combination with "formula definition".}

\item{varcomps[\code{list}].}{If not \code{NULL} passed directly to 
function \link[NVI_dynamic_linear_models]{nvi_dlm_design}, together 
with \code{systemform} and \code{errorform}. 
See \link[NVI_dynamic_linear_models]{nvi_dlm_design} 
for details.}
}
\value{
A list with elements:\cr
\item at = prior means (\code{N x p}) for parameter vector \eqn{\theta},\cr
\item Rt = array of size (\code{N x p x p}), with prior variances for parameter vector \eqn{\theta},\cr
\item mt = prior means (\code{N x p}) for parameter vector \eqn{\theta},\cr
\item Ct = array of size (\code{N x p x p}), with posterior variances for parameter vector \eqn{\theta},\cr
\item ft = forecasts for observations matrix (\code{N x q}),\cr
\item QT = array of size (\code{N x q x q}), with forecast variances.\cr
\item At = array of size (\code{N x p x q}), with AT matrices...,\cr
\item et = forecast errors (\code{N x p}),
\tem log_like = marginal log liikelihoods for each timestep (length \code{N}).\cr
\item Qtsmooth = NULL, "smoothed forcast variances", only used if \code{smoother_run = TRUE}.\cr
\item atsmooth = NULL, "smoothed forcasts", only used if \code{smoother_run = TRUE}.\cr
\item Rtsmooth = NULL, "smoothed R matrices", only used if \code{smoother_run = TRUE}.
}
\description{
The function \code{nvi_dlm} fits a Dynamic Linear Model (DLM) 
to the data input. Either the three arguments \code{systemform}, 
\code{errorform}, and \code{varcomps} are used (sent directly to 
\link[NVI_dynamic_linear_models]{nvi_dlm_design}) for constructing
design matrices or the design matrices (\code{G,F,W and V}) are given as 
direct arguments.
}
\details{
The code is under development. The overall goal is to get a more 
generic framework in the "formulas", in line with formulas used in 
\link[stats]{lm} functions and functions like \link[lme4]{lmer}.
}
\examples{
\dontrun{

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
