% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvi_dlm_design.R
\name{nvi_dlm_design}
\alias{nvi_dlm_design}
\title{nvi_dlm_design: Construct System matrix (G) and Design matrix (F)}
\usage{
nvi_dlm_design(
  systemform = formula(. ~ 0 + harmonic(waves) + nested(a \%in\% b) +
    matern(coordinates)),
  errorform = formula(. ~ 1 + b + a + ... + matern(coordinates_err)),
  varcomps = list(main = 1, trend = 10^-3, error = list(main = 1, a_err = 0.1, b_err =
    0.1), waves = 10^-3, a = NULL, b = NULL, coordinates = list(kappa = 10, nu = 1/5),
    coordinates_err = list(kappa = 10, nu = 1/5)),
  data = list(q = 1, y = NULL, a = NULL, b = NULL, a_err = NULL, b_err = NULL, waves =
    c(2 * pi/365, 2 * 2 * pi/365), coordinates = NULL, coordinates_err = NULL),
  isnullw = FALSE
)
}
\arguments{
\item{systemform}{[\code{formula}]. A object of class "formula", see
\link[stats]{formula}. The \code{formula} must be of the 
following form:\cr
\code{y ~ 1 + harmonic(wave_name) + nested(a \%in\% b) + matern(coordinates)}, 
where the different elements are:\cr
\item \code{y} (response). Name of the response variable. Might be any valid
variable name. Has to be represented in \code{data} as a \code{n x q} matrix.\cr
\item \code{1} or \code{0}. If \code{1} one common trend component is added 
to the DLM model.\cr 
\item harmonic(wave_name). Might be neglected. If part of the formula, 
harmonic waves with wave length defined in a vector (any length >=1) given 
in \code{data}.The name of the variable in data is the "\code{wave_name}" in the 
\code{harmonic(wave_name)} term. One common variance component defined in
the input parameter \code{varcomps} with name given by the argument 
("\code{wave_name}") in the formula term \code{harmonic(wave_name)} constitutes 
the elements on the main diagonal of associated block of system variance 
\code{W}. If the argument is given as "\code{1|wave_name}", somewhat similar to 
the \code{lme4} package, there will be unique wave parameters for each response
variable. 
\item nested(a\%in\%b). Might be neglected. If part of the formula, two nested
factors, where factor "\code{a}" is nested in factor "\code{b}" might be added
to the model. The factors with names \code{a} and \code{b}, any valid names 
might be chosen as long as they are part of the argument in formula term 
\code{nested(a\%in\%b)} must be present in  \code{data}, both with length 
\code{p}, i.e. the number of observations per time step. Two variance 
components defined in the input parameter \code{varcomps} with names given 
by the factors of the argument in formula ("\code{nested(a\%in\%b)}") constitutes 
the elements on the main diagonals of associated blocks of system variance 
\code{W}. 
\item matern(coordinates). Might be neglected. If part of the formula, a matèrn 
covariance structure is added to the system covariance, \code{W}. A variable 
with name defined by the argument in formula term \code{matern(coordinates)} must
be present in data. As by now this variable has to be a matrix of size 
\code{p x 2} with longitude and latitude as it is passed directly to 
\link[sp]{:pDists} with argument "\code{longlat=TRUE}". The distance matrix
(\code{p x p}) resulting from this command is passed to command
\link[rSPDE]{matern.covariance}, with parameters 
\code{kappa = varcomps[[coordinates]]$kappa}, \code{nu = varcomps[[coordinates]]$nu}, 
and \code{sigma = varcomps$main}), where \code{coordinates} 
is the argument from formula term \code{matern(coordinates)}.}

\item{errorform}{[\code{formula}]. A object of class "formula", see 
\link[stats]{formula}. The \code{formula} must be of the 
following form:\cr
\code{.~ 1 +b + a +`...` + matern(coordinates_err)}, 
where the different elements are:\cr
\item \code{1} or \code{0}. If \code{1} one common trend component is added 
to the DLM model. Might be combined with factors. \cr
\item \code{a}, \code{b}.... Might be neglected. Factors for which blocks within the error 
covariance (\code{V}) shall have equal covariance elements identified by
\code{varcomps$error[[a]]}, \code{varcomps$error[[b]]} etc. 
\item matern(coordinates_err). Might be neglected. If part of the formula, a matèrn 
covariance structure is added to the error covariance, \code{V}. A variable 
with name defined by the argument in formula term \code{matern(coordinates_err)} must
be present in data. As by now this variable has to be a matrix of size 
\code{p x 2} with longitude and latitude as it is passed directly to 
\link[sp]{:pDists} with argument "\code{longlat=TRUE}". The distance matrix
(\code{p x p}) resulting from this command is passed to command
\link[rSPDE]{matern.covariance}, with additional parameters 
\code{kappa = varcomps[[coordinates_err]]$kappa}, 
\code{nu = varcomps[[coordinates_err]]$nu}, 
and \code{sigma = varcomps$error$main}). Separate or common coordinate matrices
and parameters for \code{nu} and \code{kappa} might be applied to the error 
and system equations/ formulas.}

\item{data}{[\code{list}] with elements:\cr
\item q, [\code{integer}]. Giving the dimension of data if not defined in 
\code{systemform} (respnce) with associated data matrix \code{y}
\item y [\code{numeric matrix}] of size \code{n x q}, with data.
\item a, b, a_err,b_err,... factors of length \code{q} used in \code{systemform}
 and/ or \code{errorform}.
 \item waves [\code{numeric vector}] with wave lengths for harmonic waves.
 \item coordinates and/ or (might be the same) coordinates_err 
 [\code{numeric matrices}] giving longitude and latitude if maternal elements 
 are to be incorporated in system- (\code{W}) or error variance (\code{V}) 
 matrices.}

\item{isnullw}{[\code{logical}]. If the system variance is to be returned as 
\code{NULL}. Might be used if discount factor method is to be applied.}

\item{varcomps[\code{list}]}{with elements: \cr
\item main. Must be included with exact name "main". Common system variance.
\item trend. Must be included with exact name "trend" if trend is included 
in system variance, i.e. \code{systemform = formula(. ~ 1 + ...+)}
\item error [\code{list}] with elements \code{main} (compulsory if common 
variance component is added in error variance, i.e. 
\code{errorform = formula(. ~ 1 + ...+)}). The main error might be a vector 
of length \code{p} which gives the main diagonal of the error variance matrix
(\code{V}). Furthermore,  elements \code{a_err}, \code{b_err},...
giving variances associated with blocks defined by factors of errorform 
elements named \code{a_err}, \code{b_err}, ... 
\item waves, one scalar giving variance component in system variance 
(\code{W}) associated with harmonic waves. 
\item a, b, scalars defining variance component for the blocks in the 
system variance (\code{W}) associated with nested factors is system equation. 
Names must be part of the argument in element \code{nested(a \%in\% b)} in 
\code{systemform}.
\item a_err, b_err, ..., scalars defining variance component for the blocks in the 
error variance (\code{V}) associated with factors defined as elements in 
elements of the \code{errorform}.
\item coordinates = [\code{list}] with elements \code{kappa} and 
\code{nu} used for the matèrn covariance in system variance (\code{W}). 
\item coordinates_err = [\code{list}] with elements \code{kappa_err} and 
\code{nu_err} used for the matèrn covariance in error variance (\code{V}).}
}
\value{
A list with the following elements:\cr 
\item GG: System design (square) matrix G with dimension [q + trend == TRUE + 2 x number of waves]
\item FF: Observation design matrix with dimension [q + trend == TRUE + 2 x number of waves] x q 
\item W: System covariance matrix with dimension [q + trend == TRUE + 2 x number of waves]
\item V: Error covariance matrix with dimension q x q.
}
\description{
The function \code{nvi_dlm_design} constructs the 
\code{G,F, W and V} design matrices for DLM analyses. It is (by now) 
possible to include a multivariate response with common trend and 
common seasonality modeled as harmonic waves. Furthermore, for the system 
equation nested factors and matèrn covariance might be included.
}
\details{
...
}
\examples{
\dontrun{

}


}
\author{
Lars Erik Gangsei\cr
lars.erik.gangsei@vetinst.no\cr
+47 950 61 231
}
